moved plot functions from utils/geoutils to visualization/matplotlib
