+++ pymove/utils/geoutils.py
@@ -2,23 +2,18 @@
 Geo operations.
 
 v_color,
-plot_coords,
-plot_bounds,
-plot_line,
 create_geohash_df,
 create_bin_geohash_df,
 decode_geohash_to_latlon,
 
 """
 
-from typing import Optional, Text, Tuple, Union
+from typing import Optional, Text, Tuple
 
 import geohash2 as gh
 import numpy as np
-from matplotlib.pyplot import axes
 from numpy import ndarray
 from pandas import DataFrame
-from shapely.geometry import LineString, MultiLineString
 from shapely.geometry.base import BaseGeometry
 
 from pymove.utils.constants import (
@@ -72,86 +67,6 @@ def v_color(ob: BaseGeometry) -> Text:
     #ffcc33 <class 'str'>
     """
     return COLORS[ob.is_simple + 33]
-
-
-def plot_coords(ax: axes, ob: BaseGeometry, color: Optional[Text] = 'r'):
-    """
-    Plot the coordinates of each point of the object in a 2D chart.
-
-    Parameters
-    ----------
-    ax : axes
-        Single axes object
-    ob : geometry object
-        Any geometric object
-    color : str, optional
-        Sets the geometric object color, by default 'r'
-
-    Example
-    -------
-    """
-    x, y = ob.xy
-    ax.plot(x, y, 'o', color=color, zorder=1)
-
-
-def plot_bounds(ax: axes, ob: Union[LineString, MultiLineString], color='b'):
-    """
-    Plot the limites of geometric object.
-
-    Parameters
-    ----------
-    ax : axes
-        Single axes object
-    ob : LineString or MultiLineString
-        Geometric object formed by lines.
-    color : str, optional
-        Sets the geometric object color, by default 'b'
-
-    Example
-    -------
-
-    """
-    x, y = zip(*list((p.x, p.y) for p in ob.boundary))
-    ax.plot(x, y, '-', color=color, zorder=1)
-
-
-def plot_line(
-    ax: axes,
-    ob: LineString,
-    color: Optional[Text] = 'r',
-    alpha: Optional[float] = 0.7,
-    linewidth: Optional[float] = 3,
-    solid_capstyle: Optional[Text] = 'round',
-    zorder: Optional[float] = 2
-):
-    """
-    Plot a LineString.
-
-    Parameters
-    ----------
-    ax : axes
-        Single axes object
-    ob : LineString
-        Sequence of points.
-    color : str, optional
-        Sets the line color, by default 'r'
-    alpha : float, optional
-        Defines the opacity of the line, by default 0.7
-    linewidth : float, optional
-        Defines the line thickness, by default 3
-    solid_capstyle : str, optional
-        Defines the style of the ends of the line, by default 'round'
-    zorder : float, optional
-        Determines the default drawing order for the axes, by default 2
-
-    Example
-    -------
-    """
-    x, y = ob.xy
-    ax.plot(
-        x, y, color=color, alpha=alpha, linewidth=linewidth,
-        solid_capstyle=solid_capstyle, zorder=zorder
-    )
 
 
 def _encode(lat: float, lon: float, precision: Optional[float] = 15) -> Text:
